development-ml/ml-app/requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.37
pydantic-settings==2.7.1
asyncpg==0.30.0
requests==2.32.3
pandas==2.2.3
scikit-learn~=1.5.2
xgboost~=2.1.2
joblib==1.4.2
schedule==1.2.2
sqlalchemy-utils==0.41.2
cryptography==44.0.2
psycopg2==2.9.10